backend/requirements.txt
flask-admin
flask-bootstrap
flask-collect
flask-cors
flask-jwt
flask-restplus
flask-script
flask-security
peewee
